requirements.txt
python-telegram-bot==21.9
transmission-rpc==7.0.11
